[update] node.cppにファイル分け、コメントの削除
