+++ ConfigParse/node.hpp
@@ -14,17 +14,4 @@ class Node {
 	int          GetTokenType();
 };
 
-Node::Node(std::string &token, int token_type)
-	: token_(token), token_type_(token_type) {}
-
-Node::~Node() {}
-
-std::string &Node::GetToken() {
-	return token_;
-}
-
-int Node::GetTokenType() {
-	return token_type_;
-}
-
 #endif
